U.S. Senate Precinct-Level Returns 2016

This dataset contains precinct-level returns for elections to the U.S. Senate.

Each record in the dataset gives the number of votes reported from a precinct for a candidate.

senate_precincts_2016

Format

A dataframe with 37 variables:

year

Year of election.

state

State name.

state_postal

State U.S. Postal Service abbreviation (two-letter ISO 1366 code).

state_fips

Numeric state FIPS 5-2 code.

state_icpsr

Numeric ICPSR state code.

county_name

County name.

county_ansi

County ANSI code.

county_fips

Numeric FIPS 6-4 code, the concatenation of two-digit state_fips and three-digit county FIPS codes.

county_lat

County latitude (decimal degrees).

county_long

County longitude (decimal degrees).

jurisdiction

The name of the administrative jurisdiction, typically a county, as it appeared in source data.

precinct

The name of the precinct, as it appeared in source data.

office

The office for which the candidate ran.

district

District associated with the office, where applicable.

stage

The electoral stage, either gen for general elections or pri for primary elections.

special

Whether the election was a special election, either TRUE for special elections or FALSE otherwise.

candidate

The name of the candidate. Candidate names are standardized across jurisdictions.

candidate_last

Candidate's last name.

candidate_first

Candidate's first name.

candidate_middle

Candidate's middle name.

candidate_normalized

A normalizing transformation of the candidate's name for joins by name. This is a single word from candidate, usually the last name, in lowercase. In the case of hyphenated names, only the final name is included.

candidate_full

Candidate's official full name.

candidate_suffix

Candidate name suffix.

candidate_nickname

Candidate's nickname.

candidate_fec

Candidate's [FEC identifier](https://www.fec.gov/data/advanced/?tab=candidates). Multiple FEC IDs appear concatenated, separated by a semicolon and space (; ).

candidate_fec_name

Candidate's name as it appears in FEC data.

candidate_google

Candidate's [Google Knowledge Graph](https://developers.google.com/knowledge-graph) entity identifier.

candidate_govtrack

Candidate's [GovTrack.us](https://www.govtrack.us) identifier.

candidate_icpsr

Candidate's [ICPSR](https://www.icpsr.umich.edu) identifier.

candidate_maplight

Candidate's [MapLight](https://maplight.org) identifier.

candidate_opensecrets

Candidate's [OpenSecrets.org](https://www.opensecrets.org) identifier.

candidate_wikidata

Candidate's [WikiData](https://www.wikidata.org) identifier.

candidate_party

Candidate's party affiliation. Values may differ from party, which gives the candidate's party on the ballot.

writein

Whether the record describes a write-in candidate, either TRUE or FALSE.

party

Party of the candidate, where applicable. Candidates may run on multiple party lines, so to compute two-party vote shares or candidate vote totals, aggregate over party. Party names are standardized across jurisdictions.

mode

Vote mode, e.g., mail or Election Day.

votes

Number of votes received.

Source

http://dx.doi.org/10.7910/DVN/NLTQAD, version 2018-08-02.